+++ src/lib.rs
@@ -18,7 +18,6 @@ mod cursor;
 mod event_queue;
 mod event_reader;
 mod chunk_state;
-mod conditional_mutex;
 #[allow(dead_code)]
 mod dynamic_array;
 
@@ -26,29 +25,10 @@ mod dynamic_array;
 #[allow(dead_code)]
 mod dynamic_chunk;
 
-pub trait BoolType{
-    const VALUE: bool;
-}
-pub struct True{}
-impl BoolType for True{
-    const VALUE: bool = true;
-}
-pub struct False{}
-impl BoolType for False{
-    const VALUE: bool = false;
-}
-
 /// Epoch of EventQueue::start_position
 type StartPositionEpoch = crate::utils::Epoch<u32, {i32::MAX as u64}>;
 
-pub mod mpmc{
-    //! Multi-producer multi-consumer
-    //!
-    //! Lock-free reading. Write under lock.
-
-    pub use crate::event_queue::*;
-    pub use crate::event_reader::*;
-}
+pub mod mpmc;
 
 pub mod spmc{
     //! Single-producer multi-consumer.
@@ -62,5 +42,5 @@ pub mod spmc{
 }
 
 
-// #[cfg(test)]
-// mod tests;
+#[cfg(test)]
+mod tests;