+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Changed
+- `EventQueue::subscribe` -> `EventReader::new`
+
 ## 0.3.1
 ### Changed
 - Improved read performance on long runs.